requirements.txt
streamlit>=1.32.0
groq>=0.9.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0
shap>=0.44.0
plotly>=5.18.0
openai-whisper>=20231117
gtts>=2.5.0
supabase>=2.4.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
xgboost>=2.0.0